move "build target" rules to separate file
[ci skip-rust]

[ci skip-build-wheels]
